Refactor _unpacking.py into better named file
Move norm_param into fft_array.py since it is not really related.
While UniformValue cleanly fits into its own module.

stack-info: PR: https://github.com/QSTheory/fftarray/pull/166, branch: SJSeckmeyer/stack/19
