+++ fftarray/fft_array.py
@@ -16,7 +16,7 @@ from .backends.numpy import NumpyBackend
 
 from ._utils.ufuncs import binary_ufuncs, unary_ufunc
 from ._utils.formatting import fft_dim_table, format_bytes
-from ._utils.unpacking import UniformValue, norm_param
+from ._utils.uniform_value import UniformValue
 from ._utils.indexing import (
     LocFFTArrayIndexer, check_missing_dim_names,
     tuple_indexers_from_dict_or_tuple, tuple_indexers_from_mapping,
@@ -28,6 +28,18 @@ if TYPE_CHECKING:
 
 EllipsisType = TypeVar('EllipsisType')
 Space = Literal["pos", "freq"]
+
+T = TypeVar("T")
+
+def norm_param(val: Union[T, Iterable[T]], n: int, types) -> Tuple[T, ...]:
+    """
+       `val` has to be immutable.
+    """
+    if isinstance(val, types):
+        return (val,)*n
+
+    # TODO: Can we make this type check work?
+    return tuple(val) # type: ignore
 
 class FFTArray:
     """A single class implementing FFTs."""
